AI-SORFTWARE-DEV-PROJECT/MainServer/pythonServer/requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
altgraph==0.17.4
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.3.0
deepface==0.0.81
Deprecated==1.2.14
fastapi==0.104.1
filelock==3.13.1
fire==0.5.0
Flask==3.0.0
flatbuffers==23.5.26
gast==0.5.4
gdown==4.7.1
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
gTTS==2.5.1
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
keras==2.14.0
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
mtcnn==0.1.1
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
pefile==2023.2.7
Pillow==10.1.0
protobuf==4.25.0
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
PySocks==1.7.1
python-dateutil==2.8.2
python-engineio==4.8.2
python-multipart==0.0.6
python-socketio==5.11.0
pyttsx3==2.90
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
termcolor==2.3.0
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.4
urllib3==2.0.7
uvicorn==0.24.0.post1
websocket-client==1.7.0
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
